Implement CORS configuration, update Pedido entity to use Long for telefono, and enhance error handling in PedidoController

+++ src/main/java/foodstore/foodStore/entity/dto/Pedido/PedidoCreate.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public record PedidoCreate(
         Long idUsuario,
-        Integer telefono,
+        Long telefono,
         String direccion,
         String metodoPago,
         List<DetallePedidoCreate> detalles
